client-gtk: wrap and left-align messages in history
We now have a weird issue with the MessageBox:
Initially the ScrolledWindow does not use all available space, it only
grows when its content is growing (until it can't anymore and not all
content is visible anymore when scrolling.
However, when adding a multi-line message, the content will grow by
multiple lines, but the outer size of the ScrolledWindow only by a
single line. So scrolling might be required while the top half of the
screen is still empty.